\name{documenting-package}
\alias{documenting-package}
\alias{documenting}
\docType{package}
\title{
\packageTitle{documenting}
}
\description{
\packageDescription{documenting}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{documenting}
\packageIndices{documenting}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{documenting}

Maintainer: \packageMaintainer{documenting}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
